examples/local_container_links/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile ./requirements.in -o ./requirements.txt
absl-py==2.3.0
    # via -r ./requirements.in
blinker==1.9.0
    # via flask
click==8.2.1
    # via flask
flask==3.1.1
    # via -r ./requirements.in
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
redis==6.2.0
    # via -r ./requirements.in
werkzeug==3.1.3
    # via flask
